refactor(oauth): extract shared insufficient_scope serializer + SDK guard (Phase 0)
Move extractInsufficientScopeChallenge / serializeMcpError / jsonError out of
server/routes/mcp/tools.ts into server/utils/mcp-error-serialize.ts so every
route surface that can receive a 403 insufficient_scope (SEP-2350) shares one
recognizer. tools.ts re-exports them, keeping existing importers and
tools.serialize-error.test.ts green with zero churn.

Add isInsufficientScopeError(error) to sdk errors.ts (strict brand/name match,
no duck-typing on requiredScope) and export it beside isAuthError /
isUnauthorized401. Single SDK detection primitive, mirrors the host recognizer.
