+++ sdk/src/mcp-client-manager/index.ts
@@ -124,6 +124,7 @@ export {
   MCPAuthError,
   isAuthError,
   isUnauthorized401,
+  isInsufficientScopeError,
   isMCPAuthError,
 } from "./errors.js";
